Improve optimizer row estimates for partitioned tables.
Summary:
This fixes:  bugs.mysql.com/70265

This reverts the change in this diff:
	http://bazaar.launchpad.net/~mysql/mysql-server/5.6/revision/2661.835.1#sql/ha_partition.cc

It makes the behavior more like in 5.1, but instead of a fixed maximum of
10 partitions, the new variable "part_scan_max" controls this, though
it still defaults to 10.

This version also scans from the last partition to the first, as was
recommended by db-perf.

Test Plan:
I verified all the test changes do not indicate any failures of what the
original tests were intended to check.  All the changes still do scare be
a bit, but I can find no concrete reason to be worried.

Jenkins - all tests pass, including --big tests.

New mtr test added - passes.

Reviewers: santoshb, rudradevbasak, jtolmer

Reviewed By: rudradevbasak, jtolmer

CC: mysql-eng@, hfisk, yoshinori

Differential Revision: https://phabricator.fb.com/D960562
Differential Revision: https://phabricator.fb.com/D962567

Task ID: 2841698, 2853120