Improve optimizer row estimates for partitioned tables.
Summary:
This fixes:  bugs.mysql.com/70265

This reverts the change in this diff:
	http://bazaar.launchpad.net/~mysql/mysql-server/5.6/revision/2661.835.1#sql/ha_partition.cc

It makes the behavior more like in 5.1, but instead of a fixed maximum of
10 partitions, the new variable "partition_scan_max" controls this, though
it still defaults to 10.

This version also scans from the last partition to the first, as was
recommended by db-perf.

Test Plan:
I verified all the test changes do not indicated a failure of what the
original test was intended to check.  All the changes stlil do scare be
a bit, but I can find no concrete reason to be worried.

Jenkins - all tests pass, including --big tests.

New mtr test added - passes.

Reviewers: santoshb, rudradevbasak

Reviewed By: rudradevbasak

CC: mysql-eng@, hfisk, yoshinori

Differential Revision: https://phabricator.fb.com/D960562

Task ID: 2841698