Use read_only instead of gtid_deployment_step
Summary:
Slaves should not generate GTIDs on it's own in the
slave binlog. This can break replication with GTID protocol
when the slave which generated it's own GTIDs promoted as the master.
All slaves in the replication tier can fail with SQL_thread error
1236 after promotion with message "The slave is connecting using CHANGE MASTER
TO MASTER_AUTO_POSITION = 1, but the master has purged binary logs
containing GTIDs that the slave requires".

To avoid this slaves should not generate it's own GTIDs but should
use GTIDs received from master. This requires turning on
gtid_deployment_step on slave always. This functionality is provided by
read_only which is 1 on slave and 0 on master. So, changing the logic to
just use read_only is better than changing the static variable
gtid_deployment_step.

Test Plan: test case id modifed to use read_only.

Reviewers: jtolmer

Reviewed By: jtolmer