Limit async IO slots per array to 256 as done by upstream.
Summary:
In 5.6 there are a few configuration variables that work together
in interesting ways:
* innodb_io_capacity - max number of dirty page writes per second to do
                      (also limits other IO operations)
* innodb_write_io_threads - number of background write IO threads

This undoes a previous change. The previous change has a bad side effect and
I am not sure it is needed in 5.6. The bad side effect is that Linux AIO initialization
might fail because the number of slots in a background IO array is much larger
than upstream for the same my.cnf values.

The change was probably useful in 5.1, but the code has change enough now that I
doubt it is needed.

In upstream MySQL each background IO thread has 256 slots so the total number of
slots is 256 * innodb_write_io_threads. When the slots are full then a background
write request blocks. Given the hashing of requests to a thread then it is likely
that one thread's request segment will get full and block before others.

In the FB patch we used max(256, innodb_io_capacity) as the number of slots per
array. So with our change the total number of slots is much larger than upstream.
This is a problem when innodb_io_capacity is large and even when innodb_io_capacity
X innodb_write_io_threads is large. That can make the Linux AIO init code fail
and disable AIO.

With flash I think we want innodb_io_capacity to be much greater than the number
of slots that Linux AIO supports.  There is a dedicated thread does page write requests,
so it isn't horrible to make it block when:
 256 * innodb_write_io_threads << innodb_capacity
In 5.1 that work was done by srv_master_thread which had to do many other things so
blocking was much more of a problem.

With this change, to get more background IO slots increase innodb_write_io_threads.

Test Plan:
mtr, linkbench

Revert Plan:

Database Impact:

Memcache Impact:

Other Notes:

EImportant:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Reviewers: steaphan

Reviewed By: steaphan

CC: db-eng@lists, mysql-eng@

Differential Revision: https://phabricator.fb.com/D971744

Task ID: 1853716