WITH_PERFSCHEMA_STORAGE_ENGINE=0 is Dumb
Summary:
In the disabling of PerfSchema, I changed the cmakedefine of
WITH_PERFSCHEMA_STORAGE_ENGINE from 1 to 0.  This define is only
used with #ifdef... so that was pointless.

It still works fine either way, but I am undoing this change here.

Test Plan:
Instrumented the code and cmake configs to confirm PS is still
completely disabled by default, and completely enabled when
cmake is passed -DWITH_PERFSCHEMA_STORAGE_ENGINE=1 - all works.

Reviewers: mcallaghan

Reviewed By: mcallaghan