Disable mysql-5.6 client's detection of the OS' charset
Summary:
The default charset of "auto" now just means "default".

This disables auto-detection.  We don't want to use utf8, even if the
OS supports it, because we already use utf8... but called "latin1",
and we don't want to corrupt that by trying to convert it.

I also ripped out the test which tests this feature.

Test Plan:
Connected with client and looked at the charset settings negotiated
by the connection, both before this change (utf8) and after (latin1).

Built a dev rpm and tested this, confirmed it fixes the issue.

Internal:
The dev rpm for Harrison to test with:

~steaphan/rpms/tests/fb-mysql-5.6.12-201309261145.dev.steaphan.x86_64.rpm

Harrison and I both confirmed this is fixed with this rpm.

Reviewers: jtolmer

Reviewed By: jtolmer

CC: mysql-eng@, hfisk

Differential Revision: https://phabricator.fb.com/D985970

Task ID: 2934248
(cherry picked from commit f0c5f3aa05ee2c010912c51dab484d9f8700f1dc)