Disable mysql-5.6 client's detection of the OS' charset
Summary:
The default charset of "auto" now just means "default".

This disables auto-detection.  We don't want to use utf8, even if the
OS supports it, because we already use utf8... but called "latin1",
and we don't want to corrupt that by trying to convert it.

I also ripped out the test which tests this feature.

Test Plan:
Connected with client and looked at the charset settings negotiated
by the connection, both before this change (utf8) and after (latin1).

Built a dev rpm and tested this, confirmed it fixes the issue.

Reviewers: jtolmer

Reviewed By: jtolmer