Use aligned memory for reduced innodb_doublewrite mode
Summary:
Align the "header" used for reduced innodb_doublewrite mode to the
size of that buffer (4096), so aio writes don't fail with O_DIRECT.

Test Plan:
We saw O_DIRECT writes fail in some cases in our initial prod testing.
We will repeat that testing with this diff.

Added a new test, which tests this in O_DIRECT.
It fails without this change... but passes with it.
The new test is exactly the same as old test, but:
"--innodb-flush-method=O_DIRECT" is added to the -master.opt file.

mysqltest.sh --debug --repeat=16 --parallel=4 \
	innodb.innodb_doublewrite \
	innodb.innodb_doublewrite_odirect \
	innodb.innodb_doublewrite \
	innodb.innodb_doublewrite_odirect

mysqltest.sh --release --repeat=16 --parallel=4 \
	innodb.innodb_doublewrite \
	innodb.innodb_doublewrite_odirect \
	innodb.innodb_doublewrite \
	innodb.innodb_doublewrite_odirect

Jenkins - 'arc unit' - all tests pass.

Reviewers: rudradevbasak

Reviewed By: rudradevbasak