WL#6797 Method for clearing session state
Summary:
As of today connection pools partitioned by user use COM_CHANGE_USER during
recycle of a connection, by pool manager which does a re-authentication which
is not needed. This WL will introduce a new RPC COM_RESET_CONNECTION which
will only do the cleanup of the session context without re-authentication.

WL6797: Method for clearing session state.

Probelm: main.mysql_upgrade tests fails on rhel5-x86-64bit-valgrind platform.
Fix: As part of WL6797 in main.mysql i have introduced a testcase which does not
do a cleanup of a table wl6797.t1, because of which there is a result content
mismatch in mysqlcheck.result file. The issue is fixed by doing a proper cleanup.
This is a test file issue.

Bug#17653288 MYSQL_RESET_CONNECTION DOES NOT RESET LAST_INSERT_ID

Problem: mysql_reset_connection as part of this resetting session state
it should reset last_insert_id value which it does not.
Fix: Fix is to reset all the variables associated with last_insert_id
 in THD class.

Differential Revision: D6368145

fbshipit-source-id: 338993f