remove assert in update_binlog_end_pos
Summary:
The callers of update_binlog_end_pos don't hold
LOCK_log when the log is not in use. So this assert is
unnecessary and breaks lots of test cases.

Internal: squash with https://phabricator.fb.com/D985295
Note 57 doesn't have any assert in that function.

Test Plan: mysqltest.sh

Reviewers: steaphan

Reviewed By: steaphan

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D1021205