Add -std=c++11 to gcc builds
Summary:
We build with this, but pass it in from our build scripts.

This should be set on all builds, so it should be done here.

Test Plan: Built with it, with this option removed from our build tools.

Reviewers: santoshb

Reviewed By: santoshb

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D1029497

Task ID: 3086855
(cherry picked from commit 9cb0ced9e8ebca39c248625439767adeaf5307ff)