Stop valgrind from complaining about uninitilialized variables.
Summary:
innodb.innodb_alter_table_defragment fails in valgrind runs. The report sort of suggesting a
corrupted page but it's not clear. For example, one of the conditional jump based on uninitalized
variables report points to page_type, which is a ulint initialized from the page. Adding debug output
will make the error disappear so it's hard to do any further analysis. Changing the word "SECOND" in
defragment statement "alter table t1 defragment index SECOND" to "second" seem to avoid the error.
It's not clear as of why this happens. Actually, in the original form, the first valgrind error happens
before the statement is executed. So this diff is only to make valgrind stop complaining, not really
solving the problem.

Test Plan: mysqltest.sh --valgrind innodb.innodb_alter_table_defragment

Reviewers: steaphan

Reviewed By: steaphan