Change to C99 mode and disable a warning incompatible with C99
Summary:
Subject says it all.  -Wdeclaration-after-statement isn't
compatible with C99 (since it's a warning about a feature now allowed in
C99), so it needs to be removed.

Test Plan:
my build that didn't like -Wdeclaration-after-statement now
builds.  also confirmed with ps that -std=c99 is on the gcc command
lines.

Reviewers: steaphan

Reviewed By: steaphan

CC: santoshb, mysql-eng@

Differential Revision: https://phabricator.fb.com/D1030863