skip sending transactions without GTIDs to slaves when auto_position is used
Summary:
skip sending transactions without GTIDs to slave.
Earlier logic just handled the case where GTIDs are generated
from the middle of the binlog. But this should be more general
than that since those transactions are really not necessary.

When a slave is running with gtid_mode=ON, some transactions
may not have GTID because of read_only=1 setting. These are
the transactions that are not received from master. When this
slave is promoted as the master, sending these events to slaves
is not necessary since they are not actually part of the replication
stream.

This type of events causes
slave sql_thread to unnecessary stop with error
"When @@SESSION.GTID_NEXT is set to a GTID, you must explicitly set it again after a COMMIT or ROLLBACK"
which is skipped by doing a start slave

This diff skips sending such transactions without GTIDs to slave.
The flag gtid_event_logged is set to false when XID_EVENT is received
which is an end event of a transaction. If some events without XIDs are
present in binlog and the next transaction doesn't contain GTIDs it will
be sent to slave causing the same error mentioned above. To avoid this slave
threads sets their gtid_next to automatic avoiding the sql_thread error.

Internal:
Slave may have such type of transactions because of wormhole
writing into blackhole, when this slave is promoted it is
unnecessary to send such type of events and causes sql_thread
to stop.

This is a port of the following diff from mysql-5.6.12

https://phabricator.fb.com/D1031354

* Changes during the port
thd->variables.gtid_next.set_undefined(); is moved inside
gtid_state->update_on_flush() in 5.6.16. So the logic is adjusted
in this diff to set gtid_next='automatic' after update_on_flush()
function call. Added a test case for the scenario with DDLs.

Test Plan: mysqltest.sh

Reviewers: jtolmer

Reviewed By: jtolmer

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D1031354
Differential Revision: https://phabricator.fb.com/D1183787