Skip sending GTID-less transactions when using auto_position
Summary:
Skip sending transactions without GTIDs to slave when auto_position is
used.

Earlier logic just handled the case where GTIDs are generated
from the middle of the binlog. But this should be more general
than that since those transactions are really not necessary.

When a slave is running with gtid_mode=ON, some transactions
may not have GTID because of read_only=1 setting. These are
the transactions that are not received from master. When this
slave is promoted as the master, sending these events to slaves
is not necessary since they are not actually part of the replication
stream.

This type of events causes
slave sql_thread to unnecessary stop with error
"When @@SESSION.GTID_NEXT is set to a GTID, you must explicitly set it again after a COMMIT or ROLLBACK"
which is skipped by doing a start slave

This diff skips sending such transactions without GTIDs to slave.
The flag gtid_event_logged is set to false when XID_EVENT is received
which is an end event of a transaction. If some events without XIDs are
present in binlog and the next transaction doesn't contain GTIDs it will
be sent to slave causing the same error mentioned above. To avoid this slave
threads sets their gtid_next to automatic avoiding the sql_thread error.

Test Plan: mysqltest.sh

Reviewers: jtolmer

Reviewed By: jtolmer