Complete the porting of D533427 to 5.6.
Summary:
@Internal
This diff completes the port of "Use a separate mutex and hash table
for per-tablespace stats" (D533427) to 5.6 by using the partitioned
hash table for table stats in page_zip_* functions. This diff should
have been a part of D610797, but it was not possible because at the
time of D610797 table-level compression statistics were not added to 5.6.
Table level statistics were added to 5.6 later by D625013.

Test Plan: mtr.

Reviewers: steaphan, rongrong

Reviewed By: rongrong

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D1037019

Blame Revision: D610797