Fix innodb_doublewrite test.
Summary:
innodb_doublewrite tests the functionality of the doublewrite
buffer in full-durability and reduced-durability mode. In order
to do this, it invokes the innodb_corrupt_doublewrite_pages.py
script which corrupts pages that were on the doublewrite buffer
on disk. The problem is that innodb accesses the files using
O_DIRECT but the python script accesses them through FS cache.
This meant that sometimes the corruption (or un-corruption)
performed by the python script was not seen by innodb. In order
to fix this problem we enforce flushing of the file by calling
flush() and fsync().

Test Plan:
Run the test repeatedly using the following command:

  ../srpm/mysqltest.sh --release --repeat=16 --parallel=4 innodb.innodb_doublewrite innodb.innodb_doublewrite_odirect innodb.innodb_doublewrite innodb.innodb_doublewrite_odirect

Reviewers: steaphan, rongrong

Reviewed By: rongrong