Preserve user stats for users without any current connections
Summary:
When the connection count for a user goes to zero, we are currently
deleting the bucket from the user connection hash, which means the
user stats for that user get lost. Stop deleting the bucket.

Test Plan:
Jenkins. Manually verified I could reproduce the bug
before the fix, and cannot after.

Reviewers: rongrong, steaphan

Reviewed By: steaphan

CC: mysql-eng@, ebergen

Differential Revision: https://phabricator.fb.com/D1053384

Task ID: 3173700