unlock ssl read_lock in case of errors
Summary:
unlock ssl read_lock in case of client errors.
In some cases if the client connection using ssl and there is some
problem in setting up the ssl, the thread should release the read_lock

Internal: squash with https://phabricator.fb.com/D864428
istvan found this when he is messing up with ssl. The stack trace
confirms a hang on LOCK_use_ssl and there was a ssl error on slave.

@Internal

Test Plan: Tested on unittestdb.frc1

Reviewers: steaphan

Reviewed By: steaphan

CC: istvan, mysql-eng@

Differential Revision: https://phabricator.fb.com/D1053490