Fix zero timeout client bug
Summary:
The millisecond timeout patch and async mysql client patch have
an interaction that causes 0 timeouts to not be infinite and, instead,
be 0.  This is not an issue in most cases at facebook since we set
timeouts universally, but in cases where it isn't set, network calls
would fail immediately.

Internal: squash with //phabricator.fb.com/D998163

Test Plan:
runtests, also verified the mysql binary failed when
connecting to remote hosts before this patch, and succeeds with it

Reviewers: santoshb

Reviewed By: santoshb

CC: steaphan, mysql-eng@

Differential Revision: https://phabricator.fb.com/D1058705