Fix the way mysql merges static libraries into another
Summary:
During the build process, MySQL merges some static libraries to
create other static libraries. In order to do this, it first
extracts all the object files (.o) from all the archive files (.a)
into a temporary directory and then it creates a new archive
file from all of the object files. It basically follows the
method suggested in this stack overflow answer:
http://stackoverflow.com/a/3821949.
The problem is that extracting object files using ar is
problematic when there are two files in the archive with the
same name. ar does not store full paths in the archive it
only stores filenames. Consider the following example:
there are two c files with the same name under different directories
(x/somefile.c y/somefile.c) their object files get stored in the same
library file (libsomething.a). Then libsomething.a gets merged
with libsomethingelse.a. Following command is run to extract
the object files in libsomething.a:

  $ ar -x libsomething.a

This extracts only the last object file (y/somefile.o) in the archive.
In order to extract both we specify the instance number for the
object file:

  $ ar -x -N 1 libsomething.a somefile.o
  $ mv somefile.o somefile-1.o
  $ ar -x -N 2 libsomething.a somefile.o
  $ mv somefile.o somefile-2.o

Then the merged archive can contain both somefile-1.o and somefile-2.o.

Internal:
I hit this issue when adding bzip because both bzip and zlib have
a file named compress.c.

Test Plan: build mysql and xtrabackup.

Reviewers: steaphan

Reviewed By: steaphan

CC: mysql-eng@, trunkagent

Differential Revision: https://phabricator.fb.com/D1060883

Task ID: 1782915