Fix intermittent failure in main.bug69471
Summary:
The test did not have an ORDER BY, so it could fail with the correct
rows being output in an order different from the result file. Add an
ORDER BY.

Test Plan:
Ran main.bug69471 with the ORDER BY, but without the
associated fix in the optimizer and verified that the test failed.

Reviewers: steaphan

Reviewed By: steaphan