Optimize zlib for a non-sliding window
Summary:
Several small changes which improve zlib performance by roughly 6%

Internal:
This is a port of D1052321 to the embedded zlib version.
This additionally fixes deflate for the non-default compression levels
and strategies. This was broken in the earlier diff, and caused tests to
fail when Nizam tried to port it.

Also, now I have much better testdata to test with. The improvement dropped
from 10% in the earlier crude tests, to 6% here.

Test Plan:
On data taken from mysqldumps, converted to binary, and packed into 16K
chunks, this patch shows a time improvement of 6% over the standard.

Reviewers: steaphan, nizamordulu

Reviewed By: steaphan

CC: dcapra, mysql-eng@

Differential Revision: https://phabricator.fb.com/D1096626

Task ID: 3174396