Adapted some of Intel's patches to zlib
Summary:
Most of the changes in here were inspired by the patches described here:
http://mail.madler.net/pipermail/zlib-devel_madler.net/2013-November/003087.html

Changes in this diff include:
1) By default, use CRC32 as the hash, computed using SSE4.2
2) Also provide a better fallback hash in case of no SSE
3) By default use double-byte comparisons by enabling UNALIGNED_OK
4) Use best_len_minus_1 instead of best_len in longest_match()
5) Minor optimization in deflate_slow()

Test Plan:
These changes show a 17% improvement in compression compared to without this change.
Along with the optimizations for non-sliding window, this results in a
22% improvement.

Also checked correctness by decompressing each compressed block and verifying that it matches the original data.

Reviewers: steaphan

Reviewed By: steaphan

CC: dcapra, mysql-eng@

Differential Revision: https://phabricator.fb.com/D1101627

Task ID: 3272093