slave transaction retries for MTS
Summary:
Coordinator thread enqueus events read from the relay log to the slave
worker's jobs queue. Worker thread dequeus each element, executes the
event and deletes the event. This makes slave_transaction_retries do not
work with MTS, since there is no way of knowing all the events that were
executed in the current group. The worker thread functionality is
changed so that it dequeues only when the current transaction is
successfully committed. Worker thread just reads events starting from the
head of the jobs queue storing an index which is increased on every read
from the queue. The index is reset to 0 when worker thread encounters a
temporary error or when the current transaction is successfully
committed.

slave_transaction_retries is not supported in scenarios where
coordinator thread waits for worker thread to release some memory issues
(either if queue is full or if all the worker queues are using
slave_pending_jobs_size_max memory). In these cases executed events in
the current group are dequeued to signal coordinator thread.

In MTS row query event is deleted after the execution of row event in
cleanup_context() but the event pointer is used after the transaction is
completed by worker thread.

Fix this by not releasing row query event after the execution of row event *only* for
slave worker threads. Now worker threads delete the row query events at the end of the
transaction.

Test Plan:
Simulated lock wait timeout error in slave and verified transactions are
retried by the worker thread. rpl_parallel_conf_limits tests MTS
execution when coordinator waits for worker thread due to memory issues
(This test verifies the second scenario where slave_transaction_retries
is not supported).

Reviewers: steaphan, jtolmer

Reviewed By: steaphan, jtolmer