Add admin TCP/IP port in mysql
Summary:
This diff adds admin TCP/IP port in MySQL. This admin port is a separate
TCP/IP port speaks MySQL protocol. Only users have SUPER_ACL can make
connections to this port. Internally this port is handled by separate
thread which is not part of the main thread pool thus it can accept a
connection even when the number of connections have exceeded the limit
on the master port so it won't have starving issue when server is busy.
The thread that handles the admin port works in serial manner, i.e., it
accepts a connection request, process all the commands, and it will
return when the connection is disconnected and start accepting the next
connection request. Clients cannot make more than one connections to the
port in parallel.

A new MySQL option is introduced to set the port number for this admin
port.  e.g. --admin-port=14000, this will set the admin port at 14000.

Test Plan:
Create new unit test cases admin_connection and admin_connectiion_ipv6

Reviewers: steaphan, santoshb, rudradevbasak, txchen, jtolmer, ps

Reviewed By: steaphan, jtolmer, ps, txchen

CC: maykov, ps, mysql-eng@, ekanshpreet

Differential Revision: https://phabricator.fb.com/D1010939
Differential Revision: https://phabricator.fb.com/D1119883
Differential Revision: https://phabricator.fb.com/D1121717
Differential Revision: https://phabricator.fb.com/D1126019
Differential Revision: https://phabricator.fb.com/D1135352

Task ID: 1100118, 3496100