Protect against writing to nullptr in net_write_buff
Summary:
There are reports of segfaults from mysql trying to write when
there is no valid write buffer (typically after an aborted connect,
apparently).  This protects against writing to net->write_buf when it
isn't valid.

It's not fully clear if this is a 5.6 issue or an issue with the async
patch.  The codepaths using it seem to trigger only in very rare cases
and are inside of conenction pooling code that retries connections using
the same MYSQL object, which is arguably invalid to begin with.

Test Plan: mtr

Reviewers: santoshb, jtolmer

Reviewed By: santoshb

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D1117897

Task ID: 3395398
(cherry picked from commit 2feac73fd116ceb9774f31dfbdd40b6440c41f73)