Allow sql_slave_skip_counter if gtid_executed is empty
Summary:
Allow the usage of sql_slave_skip_counter if gtid_executed is empty.
When gtids are enabled on a slave during the rollout it may be necessary
to use sql_slave_skip_counter to skip statements manually after sql_thread
hit an error.

With gtid_mode=ON, the usage of sql_slave_skip_counter is blocked to avoid
inconsistencies in gtid_executed set, but it can be used if gtids are partially
deployed i,e; when gtid_executed is empty.

Without the use of sql_slave_skip_counter it is very difficult to
skip statements (mostly DDL) that causes sql_thread errors.

Test Plan:
verfied sql_slave_skip_counter works when gtid_mode is ON
and gtid_executed is empty on a test slave.

Reviewers: jtolmer

Reviewed By: jtolmer