Implement new compression for zlib.
Summary:
This revision adds the ability to serialize the records on a page
to a buffer before compressing the buffer. Currently only zlib
can be used to compress the serialized buffer but I will add more
compression libraries in the future. I added some new versions
of the existing stress tests that use new compression. I did not
create a new stress test for all the stress tests as I am wary of
exploding the number of stress tests.

Test Plan: mtr, new stress tests.

Reviewers: rongrong

Reviewed By: rongrong