Add binlog_disk_reads counter in user_stats
Summary:
binlog_disk_reads column is incremented per disk read mysql_binlog_send.  We hook onto calls to read_function funcptr in
IOCACHE structure to detect when a real read (as opposed to a cache hit) has occurred.

Internal:

This is a port of https://phabricator.fb.com/D567213.  The port is quite clean, some functions have been moved around to new files, but the structure of the functinons affected by the diff is basically the same. I had to edit the results of (an apparently new) test funcs_1/is_columns_is.result since it dumps the USER_STATISTICS table. This is a fairly big diff chunk because it includes the column ordinals which are shift one index for each column after where the new col is inserted.  Seems like a safe port.

arc unit -  http://ci-builds.fb.com/view/MySQL/job/mysql-precommit/1298/

Test Plan:
There is a basic test added to the suite that verifies this counter works as intended.

Reviewers: jtolmer, santoshb

Reviewed By: jtolmer

CC: db-eng@lists, mysql-eng@

Differential Revision: https://phabricator.fb.com/D1193335

Task ID: 1821456