Add binlog_disk_reads counter in user_stats
Summary:
binlog_disk_reads column is incremented per disk read mysql_binlog_send.  We hook onto calls to read_function funcptr in
IOCACHE structure to detect when a real read (as opposed to a cache hit) has occurred.

Test Plan:
There is a basic test added to the suite that verifies this counter works as intended.

Reviewers: jtolmer, santoshb

Reviewed By: jtolmer