fix a deadlock caused by btr_defragment_n_pages
Summary:
A possible deadlock situation will happen when btr_defragment tries to
free a page that's the last page in the defragmentation sequence because
it's holding the file space lock and waiting for x_lock on the next block
to update it's previous page no. We should aquire the x_lock on this block
together at the beginning of defragmentation when we aquire x_locks for all
the other blocks.

Test Plan: jenkins nightly stress tests. Hopefully nothing fails from now on.

Reviewers: nizamordulu, steaphan, jtolmer

Reviewed By: nizamordulu

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D1217943

Task ID: 3902190