Store transaction ids and rollback pointers in a mini hash table
Summary:
This diff introduces the mini hashtable which is used to store trx ids and rbps
of the records on the page.

The compaction of this hash table will come at a later diff. In the future diffs the
hash table will be compacted as transactions are committed and we no longer need
the rollback pointers.

Currently the size of the hash table is around 100 bytes plus 15 bytes for each entry.
In comparison to the current case, this means we are increasing the per-record overhead
from 15 bytes to 17 bytes in the worst case.

Test Plan: mtr

Reviewers: herman, nizamordulu, jtolmer

Reviewed By: nizamordulu

CC: mysql-eng@

Differential Revision: https://phabricator.fb.com/D1091473

Task ID: 1782915