Port the Async MySQL client to 5.6
Summary:
This diff builds loosely upon the 5.1 client work and brings async to
5.6.  It is not fully feature complete (no SSL. no compression, only
supports the default native auth) but it is sufficient for our needs and
we can extend it.

This breaks the API slightly by cleaning it up and making it more
succinct.  Code using the old async API will need to change in minor
ways (though a compatibility layer could be introduced).

This also adds a #define to indicate the nonblocking client is available.

Test Plan: mtr, testing against internal tools

Reviewers: santoshb, steaphan

Reviewed By: santoshb, steaphan

CC: jtolmer, anca, lins, mysql-eng@, mcallaghan

Differential Revision: https://phabricator.fb.com/D998163
Differential Revision: https://phabricator.fb.com/D1058705
Differential Revision: https://phabricator.fb.com/D1059524
Differential Revision: https://phabricator.fb.com/D1055577
Differential Revision: https://phabricator.fb.com/D1093841
Differential Revision: https://phabricator.fb.com/D1194937

Task ID: 1817842, 2608023, 3367559