Port v5.1 Extra Stats: SHOW ENGINE INNODB TRANSACTION STATUS
Summary:
Add SHOW ENGINE InnoDB TRANSACTION STATUS.

Not all of the locking information is available in
information_schema.innodb_trx so I put in
SHOW InnoDB TRANSACTION STATUS like in 5.0.84.

I also updated the partition_innodb test to use the new
command (it checks some locking information).

Perfschema hard-coded the total number of classes into its test results
files, so they all then fail, since this adds a class of statement.
This updates that hard-coded number in all of these results files.

Internal:
Updated for 5.6.6: Output of some new perfschema tests.

This is a full port of these v5.1 patches:
	https://phabricator.fb.com/D94547
	https://phabricator.fb.com/D94757

This code is just a direct code port, matching the changes in 5.6
to the other SHOW STATUS commands.

The depricated SHOW InnoDB TRANSACTION STATUS was not included,
since the similar forms of other SHOW STATUS are not in 5.6 either.

This also uses the return value of the call to lock_print_info_summary(),
even though it should always be one in that case, to silence a gcc warning.

Test Plan: Jenkins

Reviewers: chip

Reviewed By: chip

Differential Revision: https://phabricator.fb.com/D463353
Differential Revision: https://phabricator.fb.com/D463468
Differential Revision: https://phabricator.fb.com/D637476
Differential Revision: https://phabricator.fb.com/D797871