Port v5.1 Compression: Reduce Log Records
Summary:

InnoDB logs the entire compressed page every time the page is
compressed/re-compressed in the redo log. These log records contain the
compressed data for the entire page image. Most of the time this is
unnecessary. This revision fixes this partially by eliminating the
compression log records caused by INSERTs. Note that there will still be
some log records that contain the compressed page image because an
INSERT can cause the page to split, when that happens the data on the
page is split into two and placed on two different pages. After this,
the compressed images of both of the pages are logged instead of the
INSERT log records to both pages.

Removes the unnecessary compressed page image logging that was performed
because of UPDATEs. The DELETEs do not incur any compressed page image
log records, because they are performed by just marking records as
deleted. (Later a purge operation compactifies the pages that have holes
in them, but that too doesn't cause unnecessary page image log records)
UPDATEs are performed by first marking the existing record as deleted,
then inserting a new record therefore most of the unnecessary page-image
logging was already removed.

Move page_zip_compress_write_log_no_data() and
page_zip_parse_compress_no_data() to page0zip.ic. UNIV_INLINE functions
must be put in .ic files. It doesn't cause the build to break in debug
mode, but it does so on opt mode.

When innodb_log_compressed_pages is OFF, we may run into problems during
crash recovery because of innodb_compression_level inconsistencies. Two
ways this could happen are:

1- innodb_compression_level may be changed by a DBA during runtime.
After a crash, we don't really know which pages were compressed using
what level of compression.

2- xtrabackup has no way of knowing what level was used to compress
pages in PAGE_COMPRESS_NO_DATA entries.

Page compressions happen during page reorganization too, so we also need
to log the compression level in PAGE_REORGANIZE log records. To overcome
this, we introduce a new log record called MLOG_ZIP_PAGE_REORGANIZE
which contains the compression level used during runtime.

When an INSERT triggers page-recompression (when modification log
becomes full) InnoDB logs INSERT and ZIP_PAGE_COMPRESS_NO_DATA. During
recovery while processing the INSERT record, we should not re-compress
the page even if the INSERT triggers a re-compression because the
compression level we are using might be wrong. We just wait until the
next log record to process the page.

This exposes more zlib config parameters as system vars. The most
important of these variables is innodb_zlib_wrap which makes zlib not
write adler32 checksums between compressed blocks. The default value for
wrapping is OFF now. This should save some calls to adler32 and some
space. Changes are backwards compatible. In other words it's ok to copy
an old compressed server using xtrabackup to a server with the rpm that
has this change. This is because I updated the transaction log records
to include the value for wrap and the compression strategy in addition
to compression level. This was fairly straightforward because the same
was done earlier with compression level. Prior to this change, we stored
the compression level in a byte. Now, the spare bits in that byte are
used to store values for the new options.

innodb_zlib_wrap and innodb_zlib_strategy also effect blob compression.

It's ok to change these values dynamically. this won't break xtrabackup
or crash recovery.

Needed to backport inflateReset2() from a future version of zlib in
order to reset a decompression stream without allocating memory for its
internal state and with specifying windowBits. This was necessary to
automatically determine whether a compressed page was compressed using
adler32 checksums or not. There should not be any issues in terms of
backwards compatibility with this. After this, a clean build is required
for stress tests to pass.

For some reason, zlib's inflate() reads more than just the header if
adler32 checksums were not computed. This causes a problem when a blob
that spans multiple overflow pages is decompressed. The problem exists
for the blobs that are contained within a page too, but in that case
innodb ignores the error code from zlib and assumes that the blob ends
after the first page (which is true).

Added error messages and debug assertions when a blob truncation is
detected.

Change innodb_zlib_wrap randomly in stress tests

Test Plan:
Tested on Jenkins (arc unit)

Reviewers: nizamordulu, steaphan, jtolmer

Reviewed By: nizamordulu, steaphan, jtolmer